Remove PortfolioRollHorizontal component
